backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Subarno",email = "subathanatos97@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "python-dotenv (>=1.1.1,<2.0.0)",
    "langchain (>=0.3.27,<0.4.0)",
    "langchain-google-genai (>=2.1.9,<3.0.0)",
    "pandas (>=2.3.2,<3.0.0)",
    "matplotlib (>=3.10.5,<4.0.0)",
    "google (>=3.0.0,<4.0.0)",
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "boto3 (>=1.40.16,<2.0.0)",
    "botocore (>=1.40.30,<2.0.0)",
    "chardet (>=5.2.0,<6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
